feat: add a deleteUser function
