+++ src/index.js
@@ -7,27 +7,11 @@ import { sendJsonResponse } from './utils/sendJson.js';
 dotenv.config();
 
 const PORT = process.env.PORT || 4000;
-let users = [];
 
-// Handle DELETE requests
-const handleDeleteUser = (res, userId) => {
-    if (!userId || !uuidv4.validate(userId)) {
-        return sendJsonResponse(res, 400, { message: 'Invalid userId' });
-    }
-    const userIndex = users.findIndex((u) => u.id === userId);
-    if (userIndex === -1) {
-        return sendJsonResponse(res, 404, { message: 'User not found' });
-    }
-    users.splice(userIndex, 1);
-    sendJsonResponse(res, 204, {});
-};
-
-// Create server
 const server = http.createServer((req, res) => {
     const { pathname } = url.parse(req.url, true);
     const userId = pathname.split('/')[3];
 
-    // Parse request body
     let body = '';
     req.on('data', (chunk) => {
         body += chunk.toString();
